Show timers by in-progress workflows
Show a "(runtime duration) / (last successful runtime duration)" label
by in progress workflows.

Remove the "in progress" and "completed" labels; they're redundant. The
background color clearly indicates the state.

Remove the `org/` prefix on repo names. `pachun/emma` is now just
`emma`.

Enlarge the commit title font and change the link location it takes you
to from the commit to the workflow. The workflow links to the commit.
That should be fine.